refactor: reuse SkipSameKeyValue function and add to postprocess for category key #patch
